+++ src/Entity/Submission.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Dbp\Relay\FormalizeBundle\Entity;
 
-use ApiPlatform\Metadata\ApiProperty;
 use ApiPlatform\Metadata\ApiResource;
 use ApiPlatform\Metadata\Delete;
 use ApiPlatform\Metadata\Get;
@@ -16,9 +15,7 @@ use Dbp\Relay\FormalizeBundle\Rest\SubmissionProcessor;
 use Dbp\Relay\FormalizeBundle\Rest\SubmissionProvider;
 use Doctrine\ORM\Mapping as ORM;
 use Symfony\Component\Serializer\Annotation\Groups;
-use Symfony\Component\Serializer\Attribute\Context;
 use Symfony\Component\Serializer\Attribute\Ignore;
-use Symfony\Component\Serializer\Serializer;
 
 #[ORM\Table(name: 'formalize_submissions')]
 #[ORM\Entity]
@@ -166,31 +163,9 @@ class Submission
     #[Groups(['FormalizeSubmission:output'])]
     private ?string $identifier = null;
 
-    #[ApiProperty(
-        deprecationReason: 'Use JSON object attribute \'data\' instead',
-        openapiContext: [
-            'deprecated' => true,
-        ]
-    )]
+    #[ORM\Column(name: 'data_feed_element', type: 'text')]
     #[Groups(['FormalizeSubmission:output', 'FormalizeSubmission:input'])]
     private ?string $dataFeedElement = null;
-
-    #[ApiProperty(
-        openapiContext: [
-            'type' => 'object',
-            'example' => [
-                'firstname' => 'Joni',
-                'lastname' => 'Doe',
-            ],
-        ],
-        jsonSchemaContext: [
-            'type' => 'object',
-        ]
-    )]
-    #[ORM\Column(name: 'data_feed_element', type: 'json')]
-    #[Groups(['FormalizeSubmission:output', 'FormalizeSubmission:input'])]
-    #[Context([Serializer::EMPTY_ARRAY_AS_OBJECT => true])]
-    private ?array $data = null;
 
     #[ORM\JoinColumn(name: 'form_identifier', referencedColumnName: 'identifier', onDelete: 'CASCADE')]
     #[ORM\ManyToOne(targetEntity: Form::class)]
@@ -214,30 +189,14 @@ class Submission
         $this->identifier = $identifier;
     }
 
-    /**
-     * @deprecated Use getData() instead
-     */
     public function getDataFeedElement(): ?string
     {
         return $this->dataFeedElement;
     }
 
-    /**
-     * @deprecated use setData() instead
-     */
     public function setDataFeedElement(?string $dataFeedElement): void
     {
         $this->dataFeedElement = $dataFeedElement;
-    }
-
-    public function getData(): ?array
-    {
-        return $this->data;
-    }
-
-    public function setData(?array $data): void
-    {
-        $this->data = $data;
     }
 
     public function getForm(): ?Form
@@ -271,11 +230,11 @@ class Submission
     }
 
     /**
-     * @deprecated Use getData() instead
+     * @throws \JsonException
      */
     #[Ignore]
     public function getDataFeedElementDecoded(): array
     {
-        return $this->data;
+        return json_decode($this->dataFeedElement, true, flags: JSON_THROW_ON_ERROR);
     }
 }